Data Advising

Advice

DataToolkitBase.DataAdvice — Type

DataAdvice{context, func} <: Function

DataAdvices allow for composable, highly flexible modifications of data by encapsulating a function call. They are inspired by elisp's advice system, namely the most versitile form — :around advice, and Clojure's advisors.

A DataAdvice is esentially a function wrapper, with a priority::Int attribute. The wrapped functions should be of the form:

(post::Function, action::Function, args...; kargs...) ->
  (post::Function, action::Function, args, kargs)

A three-tuple result with kargs ommited is also accepted, in which case the an empty kargs value will be automatically substituted as the fourth value.

    input=(action args kwargs)
         ┃                 ┏╸post=identity
       ╭─╂────advisor #1───╂─╮
       ╰─╂─────────────────╂─╯
       ╭─╂────advisor #2───╂─╮
       ╰─╂─────────────────╂─╯
       ╭─╂────advisor #3───╂─╮
       ╰─╂─────────────────╂─╯
         ┃                 ┃
         ▼                 ▽
action(args; kargs) ━━━━▶ post╺━━▶ result

To specify which transforms a DataAdvice should be applied to, ensure you add the relevant type parameters to your transducing function. In cases where the transducing function is not applicable, the DataAdvice will simply act as the identity function.

After all applicable DataAdvices have been applied, action(args...; kargs...) |> post is called to produce the final result.

By modifying post via rightwards-composition (i.e. post -> post ∘ extra) the transdurers compose quite nicely.

        ╭ advisor #1 ╌╌╌╌╌╌╌╌─╮
        ┆ ╭ advisor #2 ╌╌╌╌╌╮ ┆
        ┆ ┆                 ┆ ┆
input ━━┿━┿━━━▶ function ━━━┿━┿━━▶ result
        ┆ ╰╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╯ ┆
        ╰╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╯

Without a very, very good reason all advisors should behave this way. Otherwise, when post is left-composed (i.e. post -> extra ∘ post), advisors compose like so:

        ╭ advisor #1 ╌╌╌╌╌╌─╮
        ┆ ╭ advisor #2 ╌╌╌╌╌┼╌╮
        ┆ ┆                 ┆ ┆
input ━━┿━┿━━━▶ function ━━━┿━┿━━▶ result
        ╰╌┼╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╯ ┆
          ╰╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╌╯

Constructors

DataAdvice(priority::Int, f::Function)
DataAdvice(f::Function) # priority is set to 1

Examples

Logging every time a DataSet is loaded.

loggingadvisor    = DataAdvice(
    function(post::Function, f::typeof(load), loader::DataLoader, input, outtype)
        @info "Loading $(loader.data.name)"
        (post, f, (loader, input, outtype))
    end)

Automatically committing each file written.

writecommitadvisor    = DataAdvice(
    function(post::Function, f::typeof(write), writer::DataWriter{:filesystem}, output, info)
        function writecommit(result)
            run(`git add $output`)
            run(`git commit -m "update $output"`)
            result
        end
        (post ∘ writecommit, writefn, (output, info))
    end)

Advisement points

Parsing and serialisation of data sets and collections

~DataCollection~​s, ~DataSet~​s, and ~AbstractDataTransformer~​s are advised at two stages during parsing:

When calling fromspec on the Dict representation, at the start of

parsing

At the end of the fromspec function, calling identity on the object

Serialisation is performed through the tospec call, which is also advised.

The signatures of the advised function calls are as follows:

fromspec(DataCollection, spec::Dict{String, Any}; path::Union{String, Nothing})::DataCollection
identity(collection::DataCollection)::DataCollection
tospec(collection::DataCollection)::Dict

fromspec(DataSet, collection::DataCollection, name::String, spec::Dict{String, Any})::DataSet
identity(dataset::DataSet)::DataSet
tospec(dataset::DataSet)::Dict

fromspec(ADT::Type{<:AbstractDataTransformer}, dataset::DataSet, spec::Dict{String, Any})::ADT
identity(adt::AbstractDataTransformer)::AbstractDataTransformer
tospec(adt::AbstractDataTransformer)::Dict

Processing identifiers

Both the parsing of an Identifier from a string, and the serialisation of an Identifier to a string are advised. Specifically, the following function calls:

parse(Identifier, spec::AbstractString, advised=true)
string(ident::Identifier)

The data flow arrows

The reading, writing, and storage of data may all be advised. Specifically, the following function calls:

load(loader::DataLoader, datahandle, as::Type)
storage(provider::DataStorage, as::Type; write::Bool)
save(writer::DataWriter, datahandle, info)