Data Transduction

Transduction

DataToolkitBase.DataTransducer — Type

DataTransducer{context, func} <: Function

DataTransducers allow for composible, highly flexible modifications of data. They are inspired by elisp's advice system, namely the most versitile form — :around advice, and Clojure's transducers.

A DataTransducer is esentially a function wrapper, with a priority::Int attribute. The wrapped functions should be functions of the form:

(post::Function, action::Function, args...; kargs...) ->
  (post::Function, action::Function, args, kargs)

To specify which transforms a DataTransducer should be applied to, ensure you add the relevant type parameters to your transducing function. In cases where the transducing function is not applicable, the DataTransducer will simply act as the identity function.

After all applicable DataTransducers have been applied, action(args...; kargs...) |> post is called to produce the final result.

Constructors

DataTransducer(priority::Int, f::Function)
DataTransducer(f::Function) # priority is set to 1

Examples

Should you want to log every time a DataSet is loaded, one could write the following DataTransducer:

# TODO update
loggingtransducer = DataTransducer(
    function(loader::DataLoader, loadfn, (datain, outtype), kwargs)
        @info "Loading $(loader.data.name)"
        (loadfn, loader, (datain, outtype), kwargs)
    end)

Should you wish to automatically commit each write:

# TODO update
writecommittransducer = DataTransducer(
    function(writer::DataWriter{:filesystem}, writefn::typeof(write), (output, info)::Tuple{Any, Any}, kwargs)
        writecommit(writer, output::Any, info) =
          (writefn(writer, output, info); run(`git commit $output`))
        (writer, writefn, (output, info), kwargs)
    end)

Transduction points

Construction of data sets and collections

~DataCollection~​s, ~DataStore~​s, and ~AbstractDataTransformer~​s are transduced at two stages during construction:

When calling fromspec on the Dict representation, at the start of

construction

At the end of construction, calling identity on the object

The signatures of the transduced function calls are as follows:

fromspec(DataCollection, spec::Dict{String, Any}; path::Union{String, Nothing})
identity(DataCollection)

fromspec(DataSet, collection::DataCollection, name::String, spec::Dict{String, Any})
identity(DataSet)

fromspec(ADT::Type{<:AbstractDataTransformer}, dataset::DataSet, spec::Dict{String, Any})
identity(ADT::AbstractDataTransformer)

Processing identifiers

Both the parsing of an Identifier from a string, and the serialisation of an Identifier to a string are transduced. Specifically, the following function calls:

parse(Identifier, spec::AbstractString, transduced=true)
string(ident::Identifier)

The data flow arrows

The reading, writing, and storage of data may all be transduced. Specifically, the following function calls:

load(loader::DataLoader, datahandle, as::Type)
storage(provider::DataStorage, as::Type; write::Bool)
writeinfo(writer::DataWriter, datahandle, info)