Using Packages

It is entirely likely that in the course of writing a package providing a custom data transformer, one would come across packages that may be needed.

Every possibly desired package could be shoved into the list of dependences, but this is a somewhat crude approach. A more granular approach is enabled with two macros, @addpkg and @use.

Letting DataToolkitBase know about extra packages

DataToolkitBase.@addpkg — Macro

@addpkg name::Symbol uuid::String

Register the package identifed by name with UUID uuid. This package may now be used with @use $name.

All @addpkg statements should lie within a module's __init__ function.

Example

@addpkg CSV "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"

Using extra packages

DataToolkitBase.@use — Macro

@use pkg1 pkg2...

Fetch previously registered modules.

@use pkg1 loads the module pkg1 into the current scope as pkg1. Multiple packages may be loaded all at once by seperating each package name with a space.

Example

module DataToolkitExample

using DataToolkitBase
using DataFrame

function __init__()
    @addpkg CSV "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
    @addpkg DelimitedFiles "8bb1440f-4735-579b-a4ab-409b98df4dab"
end

function load(::DataLoader{:csv}, from::IOStream, ::Type{DataFrame})
    @use CSV
    result = CSV.read(from, DataFrame)
    close(from)
    result
end

function load(::DataLoader{:delimcsv}, from::IOStream, ::Type{DataFrame})
    @use DelimitedFiles
    result = DelimitedFiles.readdlm(from, ',', DataFrame)
    close(from)
    result
end

end